# [qspi_amoled] Improve drawing performance