# What does this implement/fix?

When drawing large filled shapes (or using the very common `clear()` operation), the separate function calls for each pixel drawn cause a lot of overhead.

This PR reduces the processing time for drawing filled shapes in qspi_amoled by about an order of magnitude (see benchmarks below).

Due to the clipping and rotation logic, I have opted to optimize `filled_rectangle()`, and make `horizontal_line()`  and `vertical_line()` call it with a 1 pixel rect. `horizontal_line()` is used for drawing filled circles and polygons and significantly boosts performance there as well.

I had to make the three optimized drawing functions `virtual` in display.h to render this possible. In the event that this is not acceptable, I can make a new PR that at least overrides the (already virtual) `fill()` for better full-screen fill/clear performance.

### Benchmarks

Time spent drawing graphics to the buffer in `do_update_()`, not including the time taken to send the modified buffer to the display (QSPI). Test device: LILYGO T4 S3, rotation 0.

Test | Time before | Time after PR
---- | ------------ | ------------
test_card | 1407 ms | 147 ms
full screen fill, all pixels changed | 625 ms | 77 ms
full screen fill, nothing changed | 590 ms | 43 ms
draw filled circle with 200 px radius | 332 ms | 42 ms


## Types of changes

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Other

## Test Environment

- [ ] ESP32
- [x] ESP32 IDF
- [ ] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).