@clydebarrow 2024-06-29
Rather than add this to the qspi_amoled component, it would be better implemented in a layer that could be the parent of other buffer oriented display components. Or just use LVGL.